Changed output so it is the integer value rather than float.

+++ combined/main/ICM42688/accelerometer.h
@@ -9,8 +9,8 @@ float accelerometer_set_fs(ICM42688 *imuhdl, uint8_t value);
 float accelerometer_get_odr(ICM42688 *imuhdl);
 float accelerometer_set_odr(ICM42688 *imuhdl, uint8_t value);
 
-float accelerometer_x(ICM42688 *imuhdl);
-float accelerometer_y(ICM42688 *imuhdl);
-float accelerometer_z(ICM42688 *imuhdl);
+int16_t accelerometer_x(ICM42688 *imuhdl);
+int16_t accelerometer_y(ICM42688 *imuhdl);
+int16_t accelerometer_z(ICM42688 *imuhdl);
 
 #endif
